+++ changelog.md
@@ -15,7 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Fixed
 
 - Replace `NodeId` internal `HashMap`s with indexed storage (`ca12780`)
-- Optimize binary operator connector crossing calculation (``)
+- Optimize binary operator connector crossing calculation (`c51c8d7`)
+- Remove avoidable `String` allocation in `set_selection` (``)
 
 # [Released]